Add swotGB notes page to the emulator development section
I found this page which gives some insight into the development process of swotGB, a javascript Game Boy emulator. 
I thought it might be a good addition to awesome gbdev, as it provides nice details about the development process, some of its intricacies as well as the fun that comes with it.